Make all non-abstract classes from Kununu\Elasticsearch\Query namespace final
