+++ src/Query/Criteria/Bool/MustNot.php
@@ -5,5 +5,7 @@ namespace Kununu\Elasticsearch\Query\Criteria\Bool;
 
 final class MustNot extends AbstractBoolQuery
 {
+    use BoolQueryCreateTrait;
+
     public const string OPERATOR = 'must_not';
 }
