+++ src/Query/Criteria/Bool/Should.php
@@ -5,5 +5,7 @@ namespace Kununu\Elasticsearch\Query\Criteria\Bool;
 
 final class Should extends AbstractBoolQuery
 {
+    use BoolQueryCreateTrait;
+
     public const string OPERATOR = 'should';
 }
