+++ src/Query/Criteria/Filter/GeoDistance.php
@@ -5,14 +5,14 @@ namespace Kununu\Elasticsearch\Query\Criteria\Filter;
 
 use Kununu\Elasticsearch\Query\Criteria\GeoDistanceInterface;
 
-class GeoDistance
+final class GeoDistance
 {
     public const string KEYWORD = 'geo_distance';
 
     public static function asArray(string $field, GeoDistanceInterface $value, array $options = []): array
     {
         return [
-            static::KEYWORD => array_merge(
+            self::KEYWORD => array_merge(
                 $options,
                 [
                     'distance' => $value->getDistance(),
