+++ src/Query/Criteria/Filter/GeoShape.php
@@ -5,14 +5,14 @@ namespace Kununu\Elasticsearch\Query\Criteria\Filter;
 
 use Kununu\Elasticsearch\Query\Criteria\GeoShapeInterface;
 
-class GeoShape
+final class GeoShape
 {
     public const string KEYWORD = 'geo_shape';
 
     public static function asArray(string $field, GeoShapeInterface $value, array $options = []): array
     {
         return [
-            static::KEYWORD => [
+            self::KEYWORD => [
                 $field => array_merge(
                     $options,
                     ['shape' => $value->toArray()]
