+++ src/Query/Criteria/Filter/Prefix.php
@@ -3,14 +3,14 @@ declare(strict_types=1);
 
 namespace Kununu\Elasticsearch\Query\Criteria\Filter;
 
-class Prefix
+final class Prefix
 {
     public const string KEYWORD = 'prefix';
 
     public static function asArray(string $field, $value, array $options = []): array
     {
         return [
-            static::KEYWORD => array_merge($options, [$field => $value]),
+            self::KEYWORD => array_merge($options, [$field => $value]),
         ];
     }
 }
