+++ src/Query/Criteria/Search/MatchPhrasePrefixQuery.php
@@ -3,14 +3,14 @@ declare(strict_types=1);
 
 namespace Kununu\Elasticsearch\Query\Criteria\Search;
 
-class MatchPhrasePrefixQuery
+final class MatchPhrasePrefixQuery
 {
     public const string KEYWORD = 'match_phrase_prefix';
 
     public static function asArray(array $fields, string $queryString, array $options = []): array
     {
         return [
-            static::KEYWORD => [
+            self::KEYWORD => [
                 $fields[0] => array_merge(
                     $options,
                     [
