+++ src/Query/Criteria/Search/MatchPhraseQuery.php
@@ -3,14 +3,14 @@ declare(strict_types=1);
 
 namespace Kununu\Elasticsearch\Query\Criteria\Search;
 
-class MatchPhraseQuery
+final class MatchPhraseQuery
 {
     public const string KEYWORD = 'match_phrase';
 
     public static function asArray(array $fields, string $queryString, array $options = []): array
     {
         return [
-            static::KEYWORD => [
+            self::KEYWORD => [
                 $fields[0] => array_merge(
                     $options,
                     [
